Handle-intrinsic-functions (#171)
* Handle-intrinsic-functions

* Remove-Cli-Tests-Leftovers
