+++ CHANGELOG.txt
@@ -1,9 +1,10 @@
 0.18.18:
+ - Some advance handling of intrinsic function if passed via kwargs to operations.
  - RD-4607 Improve File System Modules Support
 0.18.17: RD-4653-support-tf-vars
 0.18.16:
  - RD-4603 Update Terraform Binary Workflow
- - RD-4693 Add import resource functionality. 
+ - RD-4693 Add import resource functionality.
 0.18.15: RD-4023 Terratag
 0.18.14: RD-4372 Make sure that terraform plan only uses module nodes.
 0.18.13: Fix issue in tflint
